Refactor page templates
