+++ src/blog/blog-components/page-template/BlogPost.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
-import { Helmet } from "react-helmet";
 import moment from "moment";
+import PageHelmut from "./template-components/PageHelmut";
+import { scrollToTop } from "./templateUtils";
 
 interface Props {
   /** Title to display on post. */
@@ -9,19 +10,17 @@ interface Props {
   date: moment.Moment;
 }
 
-/** blog post skeleton. */
+/** Blog post template. */
 const BlogPost: React.FunctionComponent<Props> = ({
   title,
   date,
   children,
 }) => {
-  useEffect(() => window.scrollTo(0, 0), []);
+  useEffect(scrollToTop, []);
 
   return (
     <>
-      <Helmet>
-        <title>{title}</title>
-      </Helmet>
+      <PageHelmut title={title} />
       <div className="blog-post">
         <h1 className="blog-title">{title}</h1>
         <span className="blog-date">{moment(date).format("MM/DD/YYYY")}</span>
