+++ src/blog/blog-components/page-template/template-components/PageHelmut.tsx
@@ -1,0 +1,16 @@
+import { Helmet } from "react-helmet";
+import React from "react";
+
+interface Props {
+  title: string;
+}
+
+function PageHelmut({ title }: Props) {
+  return (
+    <Helmet>
+      <title>{title}</title>
+    </Helmet>
+  );
+}
+
+export default PageHelmut;
