+++ src/blog/blog-components/page-template/templateUtils.ts
@@ -1,0 +1,1 @@
+export const scrollToTop = () => window.scrollTo(0, 0);
