+++ src/blog/categories/bradyanderson-tech/posts/CustomHoverDomainForHerokuAppPost.tsx
@@ -1,4 +1,4 @@
-import BlogPost from "../../../blog-components/BlogPost";
+import BlogPost from "../../../blog-components/page-template/BlogPost";
 import BlogImage from "../../../blog-components/BlogImage";
 import BlogLink from "../../../blog-components/BlogLink";
 import { BlogPostProps } from "../../../../../types/Sitemap";
