+++ src/blog/categories/retro-computing/posts/WhiteningAMacintoshPost.tsx
@@ -1,4 +1,4 @@
-import BlogPost from "../../../blog-components/BlogPost";
+import BlogPost from "../../../blog-components/page-template/BlogPost";
 import BlogImage from "../../../blog-components/BlogImage";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
